+++ tableaudocumentapi/query.py
@@ -259,7 +259,7 @@ class Query(object):
         return df
     
     @staticmethod
-    def compare_diffs(wb1_filename, wb2_filename, wb1_twb_string=None, wb2_twb_string=None):
+    def compare_workbooks(wb1_filename, wb2_filename, wb1_twb_string=None, wb2_twb_string=None):
         from tableaudocumentapi import Workbook
         if wb1_twb_string:
             d1 = Workbook(twb_xml_string=wb1_twb_string).query.get_workbook_metadata_table()